refactor: move register blocks into form builder domain
